+++ python/ray/autoscaler/_private/_kubernetes/config.py
@@ -6,7 +6,7 @@ import re
 from kubernetes import client
 from kubernetes.client.rest import ApiException
 
-from ray.autoscaler._private.kubernetes import auth_api, core_api, log_prefix
+from ray.autoscaler._private._kubernetes import auth_api, core_api, log_prefix
 import ray.ray_constants as ray_constants
 
 logger = logging.getLogger(__name__)
