+++ python/ray/autoscaler/_private/_kubernetes/kubectl-rsync.sh
